requirements_local.txt
# include requirements.txt
-r requirements.txt

django-debug-toolbar==2.2

